ai-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
numpy<2.0.0
sentence-transformers==2.3.1
torch==2.2.0 --index-url https://download.pytorch.org/whl/cpu
# Dùng bản CPU cho nhẹ, sau này deploy có GPU thì đổi sau
transformers==4.37.2
pillow
python-multipart
sentencepiece
protobuf
torch
torchvision